+++ src/Services/index.js
@@ -3,6 +3,7 @@ export * as feedbackService from "./feedbackService";
 export * as localProgramImportService from "./localProgramImportService";
 export * as programService from "./programService";
 export * as runningService from "./runningService";
+export * as socialPostService from "./socialPostService";
 export * as socialService from "./socialService";
 export * as weightliftingService from "./weightliftingService";
 export * as workoutService from "./workoutService";